feat: Starting the Project entity alongside the project itself, including corrections to the Project DTO, Repository, and Enum.

+++ src/main/java/br/com/luigifalconi/taskmanager/dto/request/ProjectRequestDTO.java
@@ -1,5 +1,6 @@
 package br.com.luigifalconi.taskmanager.dto.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import br.com.luigifalconi.taskmanager.enums.StatusProject;
@@ -20,7 +21,7 @@ public class ProjectRequestDTO {
 
     private String description;
 
-    private double budget;
+    private BigDecimal budget;
 
     private LocalDate startDate;
